Move link fetching to GetLinks component
