+++ chronograf/ui/src/index.tsx
@@ -5,7 +5,6 @@ import {render} from 'react-dom'
 import {Provider} from 'react-redux'
 import {Router, Route, useRouterHistory} from 'react-router'
 import {createHistory, History} from 'history'
-import {bindActionCreators} from 'redux'
 
 import configureStore from 'src/store/configureStore'
 import {loadLocalStorage} from 'src/localStorage'
@@ -26,9 +25,9 @@ import {SourcePage, ManageSources} from 'src/sources'
 import {UserPage} from 'src/user'
 import {LogsPage} from 'src/logs'
 import NotFound from 'src/shared/components/NotFound'
+import GetLinks from 'src/shared/containers/GetLinks'
 
 // Actions
-import {getLinksAsync} from 'src/shared/actions/links'
 import {disablePresentationMode} from 'src/shared/actions/app'
 
 // Styles
@@ -65,63 +64,41 @@ window.addEventListener('keyup', event => {
   }
 })
 
-interface State {
-  ready: boolean
-}
-
-class Root extends PureComponent<{}, State> {
-  private getLinks = bindActionCreators(getLinksAsync, dispatch)
-
-  constructor(props) {
-    super(props)
-    this.state = {
-      ready: false,
-    }
-  }
-
-  public async componentDidMount() {
-    try {
-      await this.getLinks()
-      this.setState({ready: true})
-    } catch (error) {
-      console.error('Could not get links')
-    }
-  }
-
+class Root extends PureComponent {
   public render() {
-    return this.state.ready ? (
+    return (
       <Provider store={store}>
         <Router history={history}>
-          <Route component={Setup}>
-            <Route component={Signin}>
-              <Route component={App}>
-                <Route path="/" component={CheckSources}>
-                  <Route
-                    path="dashboards/:dashboardID"
-                    component={DashboardPage}
-                  />
-                  <Route path="tasks" component={TasksPage} />
-                  <Route path="tasks/new" component={TaskPage} />
-                  <Route path="sources/new" component={SourcePage} />
-                  <Route path="data-explorer" component={DataExplorerPage} />
-                  <Route path="dashboards" component={DashboardsPage} />
-                  <Route path="manage-sources" component={ManageSources} />
-                  <Route path="manage-sources/new" component={SourcePage} />
-                  <Route
-                    path="manage-sources/:id/edit"
-                    component={SourcePage}
-                  />
-                  <Route path="user/:tab" component={UserPage} />
-                  <Route path="logs" component={LogsPage} />
+          <Route component={GetLinks}>
+            <Route component={Setup}>
+              <Route component={Signin}>
+                <Route component={App}>
+                  <Route path="/" component={CheckSources}>
+                    <Route
+                      path="dashboards/:dashboardID"
+                      component={DashboardPage}
+                    />
+                    <Route path="tasks" component={TasksPage} />
+                    <Route path="tasks/new" component={TaskPage} />
+                    <Route path="sources/new" component={SourcePage} />
+                    <Route path="data-explorer" component={DataExplorerPage} />
+                    <Route path="dashboards" component={DashboardsPage} />
+                    <Route path="manage-sources" component={ManageSources} />
+                    <Route path="manage-sources/new" component={SourcePage} />
+                    <Route
+                      path="manage-sources/:id/edit"
+                      component={SourcePage}
+                    />
+                    <Route path="user/:tab" component={UserPage} />
+                    <Route path="logs" component={LogsPage} />
+                  </Route>
                 </Route>
               </Route>
             </Route>
+            <Route path="*" component={NotFound} />
           </Route>
-          <Route path="*" component={NotFound} />
         </Router>
       </Provider>
-    ) : (
-      <div className="page-spinner" />
     )
   }
 }
